CH9121 config read/write. Simple socket to write/read data to/from TCP server (#1)
It's hard to remember all the hex commands the chip has and how to run
them. Intention of this PR is to provide python class which will hide
all low level details and expose clean interface for reading and writing
chip configuration and which will be self documented. Scope is that all
instructions, required for acting as TCP client, should be implemented from this document:
https://github.com/vmyronovych/pico-eth-ch9121/blob/cf0903937100287c8937f3729d8a72050c590a45/doc/CH9121_SPCC.pdf